requirements.txt
fastapi
uvicorn
python-dotenv
PyGithub
google-generativeai
pydantic
pydantic-settings
requests
httpx
pyjwt
cryptography
flake8
bandit
tree-sitter
tree-sitter-python
tree-sitter-javascript
tree-sitter-typescript
tree-sitter-java
tree-sitter-c
tree-sitter-cpp
tree-sitter-go
tree-sitter-rust
numpy
